LayoutTests/crypto/subtle/aes-cbc-wrong-key-class-expected.txt

Test calling AES-CBC encrypt with a HMAC key.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a raw HMAC key from string literal...
Using the key to encrypt plaintext...
PASS crypto.subtle.encrypt({name: "aes-cbc", iv: iv}, key, plaintext) threw exception Error: NotSupportedError: DOM Exception 9.
PASS successfullyParsed is true

TEST COMPLETE
